Changed get, post, and any sigs, throw Exceptions, register404Route
The get(), post(), and any() functions have been changed so they accept
three parameters, the patter, callback, and options/single filter.

The Router and Route classes can now throw custom exceptions if a route
isn't found and no 404 is registered, a filter fails or doesn't exist
and if a class or method doesn't exist or is not callable.

Added a register404Route() method to specific a specific route to
return in the event of a 404. If no route is registered a RouteException
is throw.